merombsk db dong pun

+++ database/migrations/2020_06_15_155458_create_karakteristik_table.php
@@ -16,7 +16,7 @@ class CreateKarakteristikTable extends Migration
         Schema::create('karakteristik', function (Blueprint $table) {
             $table->increments('k_id')->unique();
             $table->string('k_nama');
-            $table->float('k_bobot');
+            $table->decimal('k_bobot',8,2);
         });
     }
